chore: rename `ContextMenu` files to match convention
